requirements-test.txt
pytest
pyvirtualdisplay
pillow
backports.tempfile
